+++ src/Merger/HtmlReportMerger.php
@@ -12,38 +12,49 @@ use DOMNode;
 use DOMNodeList;
 use DOMXPath;
 use Robo\Exception\TaskException;
-use Robo\Task\BaseTask;
 use RuntimeException;
 
 /**
  * Generate common HTML report
  * Class MergeHTMLReportsTask
  * @author Kerimov Asif
+ *
+ * @see \Tests\Codeception\Task\Merger\HtmlReportMergerTest
  */
 class HtmlReportMerger extends AbstractMerger
 {
     /** @var string[] */
-    protected $src = [];
-    /** @var string */
-    protected $dst;
-    /** @var int */
-    protected $countSuccess = 0;
-    /** @var int */
-    protected $countFailed = 0;
-    /** @var int */
-    protected $countSkipped = 0;
-    /** @var int */
-    protected $countIncomplete = 0;
-    /** @var bool */
-    protected $previousLibXmlUseErrors;
+    protected array $src = [];
+
+    protected string $dst = '';
 
     /**
-     * @var float
+     * @var int|float
      */
-    private $executionTimeSum = 0;
+    protected $countSuccess = 0;
+
+    /**
+     * @var int|float
+     */
+    protected $countFailed = 0;
+
+    /**
+     * @var int|float
+     */
+    protected $countSkipped = 0;
+
+    /**
+     * @var int|float
+     */
+    protected $countIncomplete = 0;
+
+    protected bool $previousLibXmlUseErrors = false;
+
+    private float $executionTimeSum = 0;
 
     /**
      * HtmlReportMerger constructor.
+     *
      * @param string[] $src - array of source reports
      */
     public function __construct(array $src = [])
@@ -62,11 +73,11 @@ class HtmlReportMerger extends AbstractMerger
         } else {
             $this->src[] = $fileName;
         }
+
         return $this;
     }
 
     /**
-     * @param string $fileName
      * @return $this|HtmlReportMerger
      */
     public function into(string $fileName): self
@@ -75,15 +86,28 @@ class HtmlReportMerger extends AbstractMerger
         return $this;
     }
 
-    public function run()
+    public function run(): void
     {
         //save initial statament and switch on use_internal_errors mode
         $this->previousLibXmlUseErrors = libxml_use_internal_errors(true);
 
-        if (!$this->dst) {
+        if ($this->dst === '' || $this->dst === '0') {
             libxml_use_internal_errors($this->previousLibXmlUseErrors);
             throw new TaskException($this, "No destination file is set. Use `->into()` method to set result HTML");
         }
+
+        $this->printTaskInfo("Remove not existing HTML reports...");
+        foreach ($this->src as $index => $item) {
+            if (!file_exists($item)) {
+                unset($this->src[$index]);
+                $this->printTaskWarning(
+                    "HTML report {$item} did not exist and was removed from merge list"
+                );
+            }
+        }
+
+        // Resetting keys
+        $this->src = array_values($this->src);
 
         $this->printTaskInfo("Merging HTML reports into {$this->dst}");
 
@@ -96,24 +120,27 @@ class HtmlReportMerger extends AbstractMerger
         if (!$nodeList) {
             throw XPathExpressionException::malformedXPath("//table");
         }
+
         $index = 0;
-        /** @var DOMNode $table */
         $table = $nodeList->item($index);
-        if (null === $table) {
+        if (!$table instanceof DOMNode) {
             throw new KeyNotFoundException('Could not find table item at pos: ' . $index);
         }
+
         //prepare reference nodes for envs
         $xpathExprRefNodes = "//div[@class='layout']/table/tr[not(@class)]";
         $refnodes = (new DOMXPath($dstHTML))->query($xpathExprRefNodes);
         if (!$refnodes) {
             throw XPathExpressionException::malformedXPath($xpathExprRefNodes);
         }
-        for ($k = 1, $kMax = count($this->src); $k < $kMax; $k++) {
+
+        for ($k = 1, $kMax = count($this->src); $k < $kMax; ++$k) {
             $src = $this->src[$k];
             if (!file_exists($src) || !is_readable($src)) {
                 $this->printTaskWarning('File did not exists or is not readable: ' . $src);
                 continue;
             }
+
             $srcHTML = new DOMDocument();
             $srcHTML->loadHTMLFile($src, LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD);
             $this->countExecutionTime($srcHTML);
@@ -122,16 +149,19 @@ class HtmlReportMerger extends AbstractMerger
             if (!$suiteNodes) {
                 throw XPathExpressionException::malformedXPath($xpathExprSuiteNodes);
             }
+
             $j = 0;
             foreach ($suiteNodes as $suiteNode) {
                 if ($suiteNode->getAttribute('class') == '') {
                     //move to next reference node
-                    $j++;
+                    ++$j;
                     if ($j > $refnodes->length - 1) {
                         break;
                     }
+
                     continue;
                 }
+
                 //insert nodes before current reference node
                 $suiteNode = $dstHTML->importNode($suiteNode, true);
                 $table->insertBefore($suiteNode, $refnodes->item($j));
@@ -168,8 +198,9 @@ class HtmlReportMerger extends AbstractMerger
         if (!$nodeList) {
             throw XPathExpressionException::malformedXPath($xpathHeadline);
         }
+
         $pregResult = preg_match(
-            '/^Codeception Results .* \((?<timesum>\d+\.\d+)s\)$/',
+            '#^Codeception Results .* \((?<timesum>\d+\.\d+)s\)$#',
             $nodeList[0]->nodeValue,
             $matches
         );
@@ -197,23 +228,26 @@ class HtmlReportMerger extends AbstractMerger
         if (!$nodeList) {
             throw XPathExpressionException::malformedXPath($xpathHeadline);
         }
+
         /** @var DOMNode $executionTimeNode */
         $executionTimeNode = $nodeList[0]->childNodes[1]->childNodes[1];
         /** @var DOMAttr $statusAttr */
         $statusNode = $nodeList[0]->childNodes[1]->childNodes[0];
         $statusAttr = $statusNode->attributes[0];
-        if (0 !== ($this->countFailed + $this->countIncomplete + $this->countSkipped)) {
+        if (0 !== $this->countFailed) {
             $statusNode->nodeValue = 'FAILED';
             $statusAttr->value = 'color: red';
         }
-        $executionTimeNode->nodeValue = " ({$this->executionTimeSum}s)";
+
+        $executionTimeNode->nodeValue = sprintf(' (%ss)', $this->executionTimeSum);
     }
 
     /**
-    * This function counts all types of tests' scenarios and writes in class members
-    * @param DOMDocument $dstFile - destination file
-    * @throws XPathExpressionException
-    */
+     * This function counts all types of tests' scenarios and writes in class members
+     *
+     * @param DOMDocument $dstFile - destination file
+     * @throws XPathExpressionException
+     */
     private function countSummary(DOMDocument $dstFile): void
     {
         $xpathExprTests = "//table/tr[contains(@class,'scenarioRow')]";
@@ -222,6 +256,7 @@ class HtmlReportMerger extends AbstractMerger
         if (!$tests) {
             throw XPathExpressionException::malformedXPath($xpathExprTests);
         }
+
         foreach ($tests as $test) {
             $class = str_replace('scenarioRow ', '', $test->getAttribute('class'));
             switch ($class) {
@@ -243,26 +278,26 @@ class HtmlReportMerger extends AbstractMerger
 
     /**
      * This function updates values in Summary block for each type of scenarios
+     *
      * @param DOMDocument $dstFile - destination file
      */
-    private function updateSummaryTable(DOMDocument $dstFile)
+    private function updateSummaryTable(DOMDocument $dstFile): void
     {
         $dstFile = new DOMXPath($dstFile);
-        $pathFor = function ($type) {
-            return "//div[@id='stepContainerSummary']//td[@class='$type']";
-        };
-        $dstFile->query($pathFor('scenarioSuccessValue'))->item(0)->nodeValue = $this->countSuccess;
-        $dstFile->query($pathFor('scenarioFailedValue'))->item(0)->nodeValue = $this->countFailed;
-        $dstFile->query($pathFor('scenarioSkippedValue'))->item(0)->nodeValue = $this->countSkipped;
-        $dstFile->query($pathFor('scenarioIncompleteValue'))->item(0)->nodeValue = $this->countIncomplete;
+        $pathFor = fn($type): string => sprintf("//div[@id='stepContainerSummary']//td[@class='%s']", $type);
+        $dstFile->query($pathFor('scenarioSuccessValue'))->item(0)->nodeValue = (string)$this->countSuccess;
+        $dstFile->query($pathFor('scenarioFailedValue'))->item(0)->nodeValue = (string)$this->countFailed;
+        $dstFile->query($pathFor('scenarioSkippedValue'))->item(0)->nodeValue = (string)$this->countSkipped;
+        $dstFile->query($pathFor('scenarioIncompleteValue'))->item(0)->nodeValue = (string)$this->countIncomplete;
     }
 
     /**
      * This function moves Summary block in the bottom of result report
+     *
      * @param $dstFile DOMDocument - destination file
      * @param $node DOMNode - parent node of Summary table
      */
-    private function moveSummaryTable(DOMDocument $dstFile, DOMNode $node)
+    private function moveSummaryTable(DOMDocument $dstFile, DOMNode $node): void
     {
         $summaryTable = (new DOMXPath($dstFile))->query("//div[@id='stepContainerSummary']")
             ->item(0)->parentNode->parentNode;
@@ -272,14 +307,13 @@ class HtmlReportMerger extends AbstractMerger
     /**
      * This function updates values in Toolbar block for each type of scenarios
      * (blue block on the left side of the report)
+     *
      * @param DOMDocument $dstFile  - destination file
      */
-    private function updateToolbarTable(DOMDocument $dstFile)
+    private function updateToolbarTable(DOMDocument $dstFile): void
     {
         $dstFile = new DOMXPath($dstFile);
-        $pathFor = static function (string $type): string {
-            return "//ul[@id='toolbar-filter']//a[@title='$type']";
-        };
+        $pathFor = static fn(string $type): string => sprintf("//ul[@id='toolbar-filter']//a[@title='%s']", $type);
         $dstFile->query($pathFor('Successful'))->item(0)->nodeValue = '✔ ' . $this->countSuccess;
         $dstFile->query($pathFor('Failed'))->item(0)->nodeValue = '✗ ' . $this->countFailed;
         $dstFile->query($pathFor('Skipped'))->item(0)->nodeValue = 'S ' . $this->countSkipped;
@@ -288,21 +322,23 @@ class HtmlReportMerger extends AbstractMerger
 
     /**
      * This function updates "+" and "-" button for viewing test steps in final report
+     *
      * @param $dstFile DOMDocument - destination file
      * @throws XPathExpressionException
      */
-    private function updateButtons(DOMDocument $dstFile)
+    private function updateButtons(DOMDocument $dstFile): void
     {
         $xpathExprNodes = "//div[@class='layout']/table/tr[contains(@class, 'scenarioRow')]";
         $nodes = (new DOMXPath($dstFile))->query($xpathExprNodes);
         if (!$nodes) {
             throw XPathExpressionException::malformedXPath($xpathExprNodes);
         }
+
         for ($i = 2; $i < $nodes->length; $i += 2) {
             $n = $i / 2 + 1;
             $p = $nodes->item($i)->childNodes->item(1)->childNodes->item(1);
             $table = $nodes->item($i + 1)->childNodes->item(1)->childNodes->item(1);
-            $p->setAttribute('onclick', "showHide('$n', this)");
+            $p->setAttribute('onclick', "showHide('{$n}', this)");
             $table->setAttribute('id', "stepContainer" . $n);
         }
     }