+++ tests/Merger/XmlReportMergerTaskTest.php
@@ -1,5 +1,6 @@
 <?php
-declare(strict_types = 1);
+
+declare(strict_types=1);
 
 namespace Tests\Codeception\Task\Merger;
 
@@ -7,8 +8,9 @@ use Codeception\Task\Merger\XmlReportMergerTask;
 use Consolidation\Log\Logger;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Output\NullOutput;
+use const Tests\Codeception\Task\TEST_PATH;
 
-class XmlReportMergerTaskTest extends TestCase
+final class XmlReportMergerTaskTest extends TestCase
 {
     public function testMergeReports(): void
     {
@@ -16,6 +18,8 @@ class XmlReportMergerTaskTest extends TestCase
         $task->setLogger(new Logger(new NullOutput()));
         $task->from(TEST_PATH . '/fixtures/result1.xml')
             ->from(TEST_PATH . '/fixtures/result2.xml')
+            // This report did not exists and we should not throw an exception
+            ->from(TEST_PATH . '/fixtures/result3.xml')
             ->into(TEST_PATH . '/result/merged.xml')
             ->run();