+++ tests/bootstrap.php
@@ -1,9 +1,18 @@
 <?php
 
+declare(strict_types=1);
+
+namespace Tests\Codeception\Task;
+
 require_once dirname(__DIR__) . '/vendor/autoload.php';
 require_once dirname(__DIR__) . '/vendor/codeception/codeception/autoload.php';
 
 use Robo\Robo;
+use RuntimeException;
+use function dirname;
+use function is_dir;
+use function mkdir;
+use function sprintf;
 
 const TEST_PATH = __DIR__;
 
@@ -12,7 +21,7 @@ if (
     && !mkdir($concurrentDirectory)
     && !is_dir($concurrentDirectory)
 ) {
-    throw new \RuntimeException(
+    throw new RuntimeException(
         sprintf('Directory "%s" was not created', $concurrentDirectory)
     );
 }